fix(Model): switch model caused app crash
