+++ uikit/src/select/index.tsx
@@ -3,7 +3,6 @@
 import * as React from 'react'
 import {
   CaretSortIcon,
-  // CheckIcon,
   ChevronDownIcon,
   ChevronUpIcon,
 } from '@radix-ui/react-icons'
